Vignette spatial bias
First commit for the vignette spatial bias. Modifications in the function to create climate data to include imputation. Modifications in the function to create mod to solve misspelling in the argument of the function (covars_df to covars). 6 new scripts: creation of tidy spatial data, plot clim data, vignette spatial bias, function for simulations, function for simulations sncf and clim sncf.